fix: reject base aliases
